requirements.txt
flask==3.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0

